move deployContractsAndTokens
