+++ .github/workflows/call.setup-deploy-and-test-local-subgraph.yml
@@ -71,8 +71,8 @@ jobs:
         working-directory: ${{ env.sdk-core-working-directory }}
 
       - name: "Deploy Framework and Tokens"
-        run: npx hardhat run scripts/runDeployContractsAndToken.ts --network localhost
-        working-directory: ${{ env.subgraph-working-directory }}
+        run: npx hardhat run ops-scripts/runDeployContractsAndToken.ts --network localhost
+        working-directory: ./packages/ethereum-contracts
 
       - name: "Prepare files for local testing"
         run: yarn prepare-local
