+++ .github/workflows/call.test-subgraph-on-previous-sdk-core-versions.yml
@@ -51,8 +51,8 @@ jobs:
         working-directory: ${{ env.sdk-core-working-directory }}
 
       - name: "Deploy Framework and Tokens"
-        run: npx hardhat run scripts/runDeployContractsAndToken.ts --network localhost
-        working-directory: ./packages/subgraph
+        run: npx hardhat run ops-scripts/runDeployContractsAndToken.ts --network localhost
+        working-directory: ${{ env.contracts-working-directory }}
 
       - name: "Prepare files for local testing"
         run: yarn prepare-local
