+++ packages/ethereum-contracts/ops-scripts/deployContractsAndToken.ts
