+++ packages/ethereum-contracts/ops-scripts/runDeployContractsAndToken.ts
