+++ packages/ethereum-contracts/utils/README.md
@@ -32,7 +32,7 @@ This is a single-file dApp for testing the MultiFlowTesterApp contract.
 
 There are a few steps to complete to get everything up and running:
 1. Open a terminal window and start a hardhat node instance: `npx hardhat node` in `packages/ethereum-contracts`
-2. Open another terminal window and deploy contracts and token: `yarn deploy-contracts-local` in `packages/subgraph`
+2. Open another terminal window and deploy contracts and token: `npx hardhat run ops-scripts/deployContractsAndToken.ts --network localhost` in `packages/subgraph`
 3. Deploy MFA contract: `npx hardhat run ops-scripts/deploy-mfa.ts` in `packages/ethereum-contracts`
 4. Build SDK-Core: `yarn build` in `packages/sdk-core` and copy `packages/sdk-core/dist/index.umd.js` into `packages/ethereum-contracts/utils/dist`.
