+++ packages/subgraph/package.json
@@ -26,7 +26,6 @@
         "posttest": "yarn testenv:stop",
         "integrity": "npx hardhat run scripts/dataIntegrity/dataIntegrityTest.ts --network",
         "check-indexing-completeness": "ts-node scripts/checkIsDeployedOnAllNetworks.ts",
-        "deploy-contracts-local": "npx hardhat run scripts/deployContractsAndToken.ts --network localhost",
         "prepare-manifest-local": "mustache config/ganache.json subgraph.template.yaml > subgraph.yaml",
         "set-network-local": "mustache config/ganache.json src/addresses.template.ts > src/addresses.ts",
         "create-local": "graph create superfluid-test --node http://localhost:8020/",
